requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pika==1.3.1
httpx==0.26.0
redis==5.0.1
redis[cluster]==5.0.1
meilisearch==0.28.1
boto3==1.34.34
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
pytest-asyncio==0.23.3
gunicorn==21.2.0
python-multipart==0.0.6
alembic==1.13.1
# scalar-fastapi-py
phonenumbers==8.13.26
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
pybreaker==1.1.0
email-validator>=2.0.0
fastapi-mail
bcrypt==3.2.2
